Update args_manager.py to disable launching default browser at startup
When launching the script entry_with_update.py  without in argument it automatically starts a default browser. But if need to run it in headless mode without launching default browser added a new flag title --disable-browser which changes the value of auto_launch to False. This is done so that can use without launching the browser when starting it in headless mode on terminal.