Let gallery start in preview mode
Gallery will start in preview mode, which shows all of the images as thumbnails and allows the user to click on them to view them in full size, this will be the logical option when user generate 1 image at time so the generated image will not be minimized for no reason after being generated.  when generated multiple images and since the user will activate it any way to see his images it is also logical to activate it after images being generated.